wip: integrate deposit monitor into `broker_level_screening` test file.
