--- bouncer/test_commands/broker_level_screening2.ts
@@ -1,4 +1,0 @@
-#!/usr/bin/env -S pnpm tsx
-import { testBrokerLevelScreening } from '../tests/broker_level_screening2';
-
-await testBrokerLevelScreening();
